+++ src/vista/Principal.java
@@ -1,0 +1,24 @@
+package vista;
+
+import java.awt.EventQueue;
+
+import control.ParaUI;
+
+public class Principal {
+	
+	/**
+	 * Launch the application.
+	 */
+	public static void main(String[] args) {
+		EventQueue.invokeLater(new Runnable() {
+			public void run() {
+				try {
+					ParaUI frame = new ParaUI();
+					frame.setVisible(true);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		});
+	}
+}
